backend/requirement.txt
aes-everywhere==1.2.10
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorclass==2.2.2
contourpy==1.2.1
coverage==7.6.0
crispy-bootstrap4==2024.1
cssselect==1.2.0
cssutils==2.9.0
cv==1.0.0
cycler==0.12.1
DateTime==5.5
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
distlib==0.3.8
distro==1.8.0
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.2
django-environ==0.11.2
django-filter==24.2
django-heroku==0.3.1
django-localflavor==4.0
django-markdownfield==0.11.0
django-markdownx==4.0.7
django-phonenumber-field==8.0.0
django-rest-paystack==2.2.2
djangorestframework==3.15.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.1
filelock==3.15.1
Flask==3.0.2
fonttools==4.51.0
git-filter-repo==2.45.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.6
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mistune==3.0.2
MouseInfo==0.1.3
mysqlclient==2.2.4
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
numpy==1.26.4
openai==1.3.3
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
paystack==1.5.0
paystackapi==2.1.2
pefile==2023.2.7
phonenumbers==8.13.40
pickleshare==0.7.5
pillow==10.3.0
pip-upgrader==1.4.15
pipenv==2024.0.1
pipreqs==0.5.0
platformdirs==4.2.2
ply==3.11
premailer==3.10.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pur==7.3.2
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycryptodomex==3.20.0
pydantic==2.5.1
pydantic_core==2.14.3
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
PyRect==0.2.0
PyScreeze==0.1.30
pyshorteners==1.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-stdnum==1.20
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
qrcode==7.4.2
redis==5.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
setuptools==70.0.0
shortuuid==1.0.13
simple_webbrowser==0.2.2
simplejson==3.19.2
six==1.16.0
slimit==0.8.1
sniffio==1.3.0
soupsieve==2.5
SQLite3-0611==0.0.1
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.27.0
terminaltables==3.1.10
tinycss2==1.2.1
tkinter-math==0.2.4
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
tree-sitter==0.20.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.24.0.post1
virtualenv==20.26.2
virtualenv2==0.0.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.6.0
wikipedia==1.4.0
win10toast==0.9
win11toast==0.34
winsdk==1.0.0b10
yagmail==0.15.293
yamail==1.0.2
yarg==0.1.9
zope.interface==6.2
